Created block out and pause menu
Created a basic blockout of the museum level with an objective marker.

In the hut level I added an objective change that triggers on level start.

Added all the UI appropriate stuff to the BP_Hazel so that it actually appears.

Created a working pause menu that can be used wherever.